% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.vars.R
\name{best.vars}
\alias{best.vars}
\title{Best variables}
\usage{
best.vars(x)
}
\arguments{
\item{x}{A list of results of `mamglm` and `mamgllvm`}
}
\value{
A vector of terms of the best model.
}
\description{
Returns variables for the best model based on AIC
}
\examples{
#load species composition and environmental data
data(capcay)
#use a subset of data in this example to reduce run time
env_assem <- capcay$env_assem[, 1:5]
freq.abs <- as.matrix(log(capcay$abund + 1))

#to fit a gaussian regression model to frequency data:
res <- mamglm(data = env_assem, y = "freq.abs", family = "gaussian")

best.vars(res)
}
\seealso{
\code{\link{mamglm}}, \code{\link{mamgllvm}}
}
