% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.formula.R
\name{make.formula}
\alias{make.formula}
\title{Utility function}
\usage{
make.formula(lhs, vars.vec, rand.vec = NULL)
}
\arguments{
\item{lhs}{Numeric vector of dependent variables.}

\item{vars.vec}{Character vector of independet variables.}

\item{rand.vec}{Character vector of random variables (default = NULL).}
}
\value{
an object of class '"formula"'
}
\description{
Utility function for data manipulation, which is implemented in \link{maglm} and \link{mamglm}.
}
\seealso{
\code{\link{maglm}}, \code{\link{mamglm}}
}
