% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mamgllvm.R
\name{mamgllvm}
\alias{mamgllvm}
\title{Model averaging for multivariate generalized linear latent variable models}
\usage{
mamgllvm(data, y, family, scale = TRUE, AIC.restricted = FALSE)
}
\arguments{
\item{data}{Data frame, typically of environmental variables. Rows for sites and colmuns for environmental variables.}

\item{y}{Name of 'site x spcies matrix' (col for species, row for sites) (character)}

\item{family}{the 'family' object used.}

\item{scale}{Whether to scale independent variables (default = TRUE)}

\item{AIC.restricted}{Whether to use AICc (TRUE) or AIC (FALSE) (default = TRUE).}
}
\value{
A list of results

\item{res.table }{data frame with "AIC", AIC of the model, "log.L", log-likelihood of the model, "delta.aic", AIC difference to the best model, "wAIC", weighted AIC to the model, "n.vars", number of variables in the model, and each term.}

\item{importance }{vector of relative importance value of each term, caluclated as as um of the weighted AIC over all of the model in whith the term aperars.}

\item{family }{the 'family' object used.}
}
\description{
Model averaging for multivariate GLLVM based on information theory.
}
\examples{
#load species composition and environmental data
library(mvabund)
data(capcay)
#use a subset of data in this example to reduce run time
env_assem <- capcay$env_assem[, 1:2]
freq.abs <- mvabund(log(capcay$abund + 1))

#to fit a gaussian regression model to frequency data:
mamgllvm(data = env_assem, y = "freq.abs", family = "gaussian")

#to fit a binomial regression model to presence/absence data"
pre.abs0 <- capcay$abund
pre.abs0[pre.abs0 > 0] = 1
pre.abs <- mvabund(pre.abs0)

mamgllvm(data = env_assem, y = "pre.abs", family = "binomial")
}
\references{
Burnham, K.P. & Anderson, D.R. (2002) Model selection and multi-model inference: a practical information-theoretic approach. Springer Verlag, New York.

Niku, J., Warton,  D. I., Hui, F. K. C., and Taskinen, S. (2017). Generalized linear latent variable models for multivariate count and biomass data in ecology. Journal of Agricultural, Biological, and Environmental Statistics, 22:498-522.

Niku, J., Brooks, W., Herliansyah, R., Hui, F. K. C., Taskinen, S., and Warton,  D. I. (2018). Efficient estimation of generalized linear latent variable models. PLoS One, 14(5):1-20.

Warton, D. I., Guillaume Blanchet, F., O'Hara, R. B., Ovaskainen, O., Taskinen, S., Walker, S. C. and Hui, F. K. C. (2015). So many variables: Joint modeling in community ecology. Trends in Ecology & Evolution, 30:766-779.
}
\seealso{
\code{\link{maglm}}, \code{\link{ses.mamglm}}, \code{\link{ses.mamgllvm}}
}
