% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mglmn.R
\name{summary.mamglm}
\alias{summary.mamglm}
\alias{summary.mglmn}
\title{Summarize Model Averaging for Multivariate Models}
\usage{
\method{summary}{mglmn}(object, top_n = 3, digits = 3)
}
\arguments{
\item{object}{mamglm or mamgllvm object}

\item{top_n}{Number of models to show}

\item{digits}{Number of digits to use for rounding.}
}
\value{
A list of results

\item{res.table }{data frame with "AIC", AIC of the model, "log.L", log-likelihood of the model, "delta.aic", AIC difference to the best model, "wAIC", weighted AIC to the model, "n.vars", number of variables in the model, and each term.}
}
\description{
The \code{summary.mamglm} provides a summary of the
}
