% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm-fun.R
\name{maglm}
\alias{maglm}
\alias{ses.maglm}
\title{Removed Functions}
\usage{
maglm(x, ...)

ses.maglm(x, ...)
}
\arguments{
\item{x, }{removed parameters}
}
\description{
`maglm()` and `ses.maglm()` were removed as of mglmn 0.2.0. Please use `MuMIn::model.ave()` for univariate analyses instead.
}
\keyword{internal}
